+++ R/nhclu_pam.R
@@ -155,17 +155,20 @@ nhclu_pam <- function(dissimilarity,
                        algorithm_in_output = algorithm_in_output,
                        ...)
   
+  # Determine data_type based on pairwise metric
+  pairwise_metric_value <- ifelse(!inherits(dissimilarity, "dist"), 
+                                   ifelse(is.numeric(index), names(net)[3], index), 
+                                   NA)
+  data_type <- detect_data_type_from_metric(pairwise_metric_value)
+  
   outputs$inputs <- list(bipartite = FALSE,
                          weight = TRUE,
                          pairwise = TRUE,
-                         pairwise_metric = ifelse(!inherits(dissimilarity, 
-                                                            "dist"), 
-                                                  ifelse(is.numeric(index), 
-                                                         names(net)[3], index), 
-                                                  NA),
+                         pairwise_metric = pairwise_metric_value,
                          dissimilarity = TRUE,
                          nb_sites = attr(dist.obj, "Size"),
-                         hierarchical = FALSE)
+                         hierarchical = FALSE,
+                         data_type = data_type)
   
   outputs$algorithm <- list()